read explanation in description
https://practice.geeksforgeeks.org/problems/next-happy-number4538/1


Explanation:

The numbers that, when you repeatedly sum the squares of their digits, eventually result in 1 are known as "happy numbers." Let's find the numbers less than 10 that satisfy this condition:

Starting with the number 1:

1^2 = 1
Since we have already reached 1, the process stops here. 1 is a happy number.
Starting with the number 2:

2^2 = 4
4^2 = 16
1^2 + 6^2 = 37
3^2 + 7^2 = 58
5^2 + 8^2 = 89
8^2 + 9^2 = 145
1^2 + 4^2 + 5^2 = 42
4^2 + 2^2 = 20
2^2 + 0^2 = 4
Since we have reached 4 again, the process will continue in an infinite loop. 2 is not a happy number.
Starting with the number 3:

3^2 = 9
9^2 = 81
8^2 + 1^2 = 65
6^2 + 5^2 = 61
6^2 + 1^2 = 37
3^2 + 7^2 = 58
5^2 + 8^2 = 89
8^2 + 9^2 = 145
1^2 + 4^2 + 5^2 = 42
4^2 + 2^2 = 20
2^2 + 0^2 = 4
Since we have reached 4 again, the process will continue in an infinite loop. 3 is not a happy number.
Starting with the number 4:

4^2 = 16
1^2 + 6^2 = 37
3^2 + 7^2 = 58
5^2 + 8^2 = 89
8^2 + 9^2 = 145
1^2 + 4^2 + 5^2 = 42
4^2 + 2^2 = 20
2^2 + 0^2 = 4
Since we have reached 4 again, the process will continue in an infinite loop. 4 is not a happy number.
Starting with the number 5:

5^2 = 25
2^2 + 5^2 = 29
2^2 + 9^2 = 85
8^2 + 5^2 = 89
8^2 + 9^2 = 145
1^2 + 4^2 + 5^2 = 42
4^2 + 2^2 = 20
2^2 + 0^2 = 4
Since we have reached 4 again, the process will continue in an infinite loop. 5 is not a happy number.
Starting with the number 6:

6^2 = 36
3^2 + 6^2 = 45
4^2 + 5^2 = 41
4^2 + 1^2 = 17
1^2 + 7^2 = 50
5^2 + 0^2 = 25
2^2 + 5^2 = 29
2^2 + 9^2 = 85
8^2 + 5^2 = 89
8^2 + 9^2 = 145
1^2 + 4^2 + 5^2 = 42
4^2 + 2^2 = 20
2^2 + 0^2 = 4
Since we have reached 4 again, the process will continue in an infinite loop. 6 is not a happy number.
Starting with the number 7:

7^2 = 49
4^2 + 9^2 = 97
9^2 + 7^2 = 130
1^2 + 3^2 + 0^2 = 10
1^2 + 0^2 = 1
Since we have reached 1, the process stops here. 7 is a happy number.
Starting with the number 8:

8^2 = 64
6^2 + 4^2 = 52
5^2 + 2^2 = 29
2^2 + 9^2 = 85
8^2 + 5^2 = 89
8^2 + 9^2 = 145
1^2 + 4^2 + 5^2 = 42
4^2 + 2^2 = 20
2^2 + 0^2 = 4
Since we have reached 4 again, the process will continue in an infinite loop. 8 is not a happy number.
Starting with the number 9:

9^2 = 81
8^2 + 1^2 = 65
6^2 + 5^2 = 61
6^2 + 1^2 = 37
3^2 + 7^2 = 58
5^2 + 8^2 = 89
8^2 + 9^2 = 145
1^2 + 4^2 + 5^2 = 42
4^2 + 2^2 = 20
2^2 + 0^2 = 4
Since we have reached 4 again, the process will continue in an infinite loop. 9 is not a happy number.
Based on this analysis, the numbers less than 10 that result in 1 when you repeatedly sum the squares of their digits are: 1 and 7.

Code:

int nextHappy(int n){
        // code here
        n++;
        while(n<=1e5){
            if(solve(n)) return n; 
            n++;
        }
        return -1;
    }
    bool solve(int n){
        if(n == 2 || n==3 || n==4 || n==5 || n==6 || n==8 || n==9) return false;
        if( n == 1) return true;
        int sum=0;
        while(n > 0){
            sum+=((n%10)*(n%10));
            n=n/10;
        }
        if(solve(sum)) return true; 
        return false;
    }